Extract part of Mediafile to dialect package
 - Extract methods that can improve speed by rewriting optimal queries
   for a specific DB.
